Directional Bird Audio Prototype
The start of the direction bird audio has been started and in working condition.

This system will get all the trees in a given area, then plays a random bird chip sound with a massive attenuation range on a random tree.

There is a delay before playing another chirp, this can be adjusted to our likely, it has been set to 5-10 seconds for testing.