Audio Settings
Did a little rework of the audio option in the settings menu. It now works with all game audio (Besides menu sounds, those are not mine but I will modify them, and a few other sounds).

Also added in a small clamp offset to prevent audio from turning off for good when you set it to 0.

Saving is also now a thing! But only in the instance of the game, if you restart it will not save.