Uploaded Human NPC textures/Models
Uploaded all of the Human NPC textures/models, they're in the textured folder, there are  folders for different body types. The hats, accessories, shoes, and hair have their own folders because they can be used for all npc body types, however the suit for the slim male model is a whole outfit it cannot be randomized with other pieces.